+++ NEWS.md
@@ -1,5 +1,7 @@
 # milQuant 2.1.0 _2026-03-XX_
 * Use idaifieldR v0.4.0
+  * Updated connection settings. 
+  * Updated usage of simplify_idaifield().
 
 # milQuant 2.0.3 _2025-08-21_
 * Fix bug in Find Quantification where Pottery_Quantification_B was counted alongside Pottery_Quantification_A